abstracted much of main.go to kruise.go
